bridge/requirements.txt
python-osc>=1.8.0
websockets>=11.0
pyserial>=3.5